feat: migrate language selector dropdown to ts
